fix issues with regexes, ( this time hopefully for good :) ) and add another format extraction test, this time testing C language
the problem was that ingrid parser module actually espaced even escapes, which is actually not necessary and it was causing more harm then good. When I stopped escaping them,
I could also remove some extra serialization code, because it turned out that it's whole reason of existence was just to fix this problem with extra backslashes.